src/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
aws-xray-sdk==2.12.1
httpx==0.26.0
pydantic==2.5.0
python-dotenv==1.0.0
